refactors seed file to be self-contained
